2018.3.5

+++ app/views/api_v1/communities/models.py
@@ -16,8 +16,8 @@ community_model=api.model('CommunityModel',{
     'save_distance':fields.Integer,
     'eva_distance':fields.Integer,
     'ins_id':fields.String,
-    'longitude':fields.Decimal,
-    'latitude':fields.Decimal,
+    'longitude':fields.Float,
+    'latitude':fields.Float,
     'location_id': fields.String,
     'community_picture':fields.Nested(community_pic_model)
 
@@ -30,8 +30,8 @@ home_model = api.model('HomeModel', {
     'detail_address': fields.String,
     'link_name': fields.String,
     'telephone': fields.String,
-    'longitude': fields.Decimal,
-    'latitude':fields.Decimal,
+    'longitude': fields.Float,
+    'latitude':fields.Float,
     'alternate_phone':fields.String
 })
