+++ app/views/api_v1/institutes/model.py
@@ -15,8 +15,8 @@ institute_model = api.model('InstituteModel', {
     'admin_user_id': fields.String,
     'location_id':fields.String,
     'note': fields.String,
-    'longitude': fields.Decimal,
-    'latitude': fields.Decimal
+    'longitude': fields.Float,
+    'latitude': fields.Float
